pyproject.toml
[project]
name = "tiny-url"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.11"
dependencies = [
    "alembic>=1.15.0",
    "bcrypt>=4.2.0",
    "fastapi>=0.115.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "redis>=5.2.0",
    "sqlalchemy>=2.0.36",
    "structlog>=24.4.0",
    "uvicorn[standard]>=0.34.0",
    "validators>=0.34.0",
]
